requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
future==1.0.0
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
serializers==0.2.4
sqlparse==0.5.2
tzdata==2024.2
gunicorn==23.0.0
dj_database_url==2.3.0
requests=2.32.3